Got messages somewhat working:
Seems that adding a tag field with serde to a struct does not mean that serde understands to not try to deserialize the value. There should be some way to ignore it on deserialization, but it might be smarter to use an enum over the message types in the future.